+++ src/main/java/nextstep/courses/domain/session/type/FreeType.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain.session.type;
 public class FreeType implements SessionType {
 
     @Override
-    public SessionType enroll(int payAmount) {
+    public SessionType enroll(long payAmount) {
         return this;
     }
 }
