+++ src/main/java/nextstep/courses/domain/session/type/SessionType.java
@@ -1,5 +1,5 @@
 package nextstep.courses.domain.session.type;
 
 public interface SessionType {
-    SessionType enroll(int payAmount);
+    SessionType enroll(long payAmount);
 }
